requirements.txt
# Core dependencies for simple AI agent patterns
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0
fireworks-ai>=0.19.0
requests>=2.25.0
python-dotenv>=1.0.0